Re-tooled rendering engine
