+++ d2common/size.go
@@ -1,0 +1,5 @@
+package d2common
+
+type Size struct {
+	Width, Height int
+}
