+++ d2core/d2map/d2mapentity/missile.go
@@ -1,4 +1,4 @@
-package d2map
+package d2mapentity
 
 import (
 	"fmt"
