+++ d2core/d2map/d2mapentity/player.go
@@ -1,4 +1,4 @@
-package d2map
+package d2mapentity
 
 import (
 	"image/color"
