+++ d2core/d2map/d2maprenderer/camera.go
@@ -1,4 +1,4 @@
-package d2map
+package d2maprenderer
 
 type Camera struct {
 	x float64
