+++ d2core/d2map/d2maprenderer/viewport.go
@@ -1,4 +1,4 @@
-package d2map
+package d2maprenderer
 
 import (
 	"math"
@@ -24,7 +24,7 @@ func NewViewport(x, y, width, height int) *Viewport {
 			Left:   x,
 			Top:    y,
 			Width:  width,
-			Height: height,
+			Height: height + 200, // TODO: Temporary hack to prevent clipping
 		},
 	}
 }
@@ -95,10 +95,11 @@ func (v *Viewport) GetTranslationScreen() (int, int) {
 	return v.OrthoToScreen(v.transCurrent.x, v.transCurrent.y)
 }
 
-func (v *Viewport) PushTranslationOrtho(x, y float64) {
+func (v *Viewport) PushTranslationOrtho(x, y float64) *Viewport {
 	v.transStack = append(v.transStack, v.transCurrent)
 	v.transCurrent.x += x
 	v.transCurrent.y += y
+	return v
 }
 
 func (v *Viewport) PushTranslationWorld(x, y float64) {
