+++ d2game/d2gamescreen/map_engine_testing.go
@@ -1,20 +1,20 @@
 package d2gamescreen
 
 import (
-	"math"
 	"math/rand"
 	"os"
 	"time"
 
-	"github.com/OpenDiablo2/OpenDiablo2/d2game/d2player"
+	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2map/d2mapgen"
+
+	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2map/d2mapengine"
+	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2map/d2maprenderer"
 
 	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2enum"
-	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2fileformats/d2dt1"
-
 	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2input"
-	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2map"
 	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2render"
 	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2screen"
+	"github.com/OpenDiablo2/OpenDiablo2/d2game/d2player"
 )
 
 type RegionSpec struct {
@@ -80,8 +80,8 @@ var regions = []RegionSpec{
 
 type MapEngineTest struct {
 	gameState   *d2player.PlayerState
-	mapEngine   *d2map.MapEngine
-	mapRenderer *d2map.MapRenderer
+	mapEngine   *d2mapengine.MapEngine
+	mapRenderer *d2maprenderer.MapRenderer
 
 	//TODO: this is region specific properties, should be refactored for multi-region rendering
 	currentRegion int
@@ -105,7 +105,7 @@ func CreateMapEngineTest(currentRegion int, levelPreset int) *MapEngineTest {
 }
 
 func (met *MapEngineTest) LoadRegionByIndex(n int, levelPreset, fileIndex int) {
-	d2map.InvalidateImageCache()
+	d2maprenderer.InvalidateImageCache()
 	for _, spec := range regions {
 		if spec.regionType == d2enum.RegionIdType(n) {
 			met.regionSpec = spec
@@ -131,20 +131,20 @@ func (met *MapEngineTest) LoadRegionByIndex(n int, levelPreset, fileIndex int) {
 
 	if n == 0 {
 		rand.Seed(time.Now().UnixNano())
-		met.mapEngine.GenerateAct1Overworld(true)
+		d2mapgen.GenerateAct1Overworld(met.mapEngine)
 	} else {
-		met.mapEngine = d2map.CreateMapEngine(0) // necessary for map name update
+		met.mapEngine = d2mapengine.CreateMapEngine() // necessary for map name update
 		met.mapEngine.GenerateMap(d2enum.RegionIdType(n), levelPreset, fileIndex, true)
-		met.mapRenderer.SetMapEngine(met.mapEngine)
+		met.mapEngine.RegenerateWalkPaths()
 	}
-
+	met.mapRenderer.SetMapEngine(met.mapEngine)
 	met.mapRenderer.MoveCameraTo(met.mapRenderer.WorldToOrtho(met.mapEngine.GetCenterPosition()))
 }
 
 func (met *MapEngineTest) OnLoad() error {
 	d2input.BindHandler(met)
-	met.mapEngine = d2map.CreateMapEngine(0)
-	met.mapRenderer = d2map.CreateMapRenderer(met.mapEngine)
+	met.mapEngine = d2mapengine.CreateMapEngine()
+	met.mapRenderer = d2maprenderer.CreateMapRenderer(met.mapEngine)
 	met.LoadRegionByIndex(met.currentRegion, met.levelPreset, met.fileIndex)
 
 	return nil
@@ -158,121 +158,113 @@ func (met *MapEngineTest) OnUnload() error {
 func (met *MapEngineTest) Render(screen d2render.Surface) error {
 	met.mapRenderer.Render(screen)
 
-	screenX, screenY := d2render.GetCursorPos()
-	worldX, worldY := met.mapRenderer.ScreenToWorld(screenX, screenY)
-	//subtileX := int(math.Ceil(math.Mod(worldX*10, 10))) / 2
-	//subtileY := int(math.Ceil(math.Mod(worldY*10, 10))) / 2
-
-	curRegion := met.mapEngine.GetRegionAtTile(int(worldX), int(worldY))
-	if curRegion == nil {
-		return nil
-	}
-
-	tileRect := curRegion.GetTileRect()
-
-	levelFilesToPick := make([]string, 0)
-	fileIndex := met.fileIndex
-	levelPreset := curRegion.GetLevelPreset()
-	regionPath := curRegion.GetPath()
-	for n, fileRecord := range levelPreset.Files {
-		if len(fileRecord) == 0 || fileRecord == "" || fileRecord == "0" {
-			continue
-		}
-		levelFilesToPick = append(levelFilesToPick, fileRecord)
-		if fileRecord == regionPath {
-			fileIndex = n
-		}
-	}
-	if met.fileIndex == -1 {
-		met.fileIndex = fileIndex
-	}
-	met.filesCount = len(levelFilesToPick)
-
-	tileX := int(math.Floor(worldX)) - tileRect.Left
-	tileY := int(math.Floor(worldY)) - tileRect.Top
-
-	subtileX := int((worldX - float64(int(worldX))) * 5)
-	subtileY := int((worldY - float64(int(worldY))) * 5)
-
-	regionWidth, regionHeight := curRegion.GetTileSize()
-	if tileX >= 0 && tileY >= 0 && tileX < regionWidth && tileY < regionHeight {
-		tile := curRegion.GetTile(tileX, tileY)
-		screen.PushTranslation(5, 5)
-		screen.DrawText("%d, %d (Tile %d.%d, %d.%d)", screenX, screenY, tileX, subtileX, tileY, subtileY)
-		screen.PushTranslation(0, 16)
-		screen.DrawText("Map: " + curRegion.GetLevelType().Name)
-		screen.PushTranslation(0, 16)
-		screen.DrawText("%v: %v/%v [%v, %v]", regionPath, fileIndex+1, met.filesCount, met.currentRegion, met.levelPreset)
-		screen.PushTranslation(0, 16)
-		screen.DrawText("N - next region, P - previous region")
-		screen.PushTranslation(0, 16)
-		screen.DrawText("Shift+N - next preset, Shift+P - previous preset")
-		screen.PushTranslation(0, 16)
-		screen.DrawText("Ctrl+N - next file, Ctrl+P - previous file")
-		screen.PushTranslation(0, 16)
-		popN := 7
-		if len(tile.Floors) > 0 {
-			screen.PushTranslation(0, 16)
-			screen.DrawText("Floors:")
-			screen.PushTranslation(16, 0)
-			for idx, floor := range tile.Floors {
-				popN++
-				screen.PushTranslation(0, 16)
-				tileData := curRegion.GetTileData(int32(floor.Style), int32(floor.Sequence), d2enum.Floor)
-				tileSubAttrs := d2dt1.SubTileFlags{}
-				if tileData != nil {
-					tileSubAttrs = *tileData.GetSubTileFlags(subtileX, subtileY)
-				}
-				screen.DrawText("Floor %v: [ANI:%t] %s", idx, floor.Animated, tileSubAttrs.DebugString())
-
-			}
-			screen.PushTranslation(-16, 0)
-			popN += 3
-		}
-		if len(tile.Walls) > 0 {
-			screen.PushTranslation(0, 16)
-			screen.DrawText("Walls:")
-			screen.PushTranslation(16, 0)
-			for idx, wall := range tile.Walls {
-				popN++
-				screen.PushTranslation(0, 16)
-				tileData := curRegion.GetTileData(int32(wall.Style), int32(wall.Sequence), d2enum.Floor)
-				tileSubAttrs := d2dt1.SubTileFlags{}
-				if tileData != nil {
-					tileSubAttrs = *tileData.GetSubTileFlags(subtileX, subtileY)
-				}
-				screen.DrawText("Wall %v: [HID:%t] %s", idx, wall.Hidden, tileSubAttrs.DebugString())
-
-			}
-			screen.PushTranslation(-16, 0)
-			popN += 3
-		}
-		if len(tile.Walls) > 0 {
-			screen.PushTranslation(0, 16)
-			screen.DrawText("Shadows:")
-			screen.PushTranslation(16, 0)
-			for idx, shadow := range tile.Shadows {
-				popN++
-				screen.PushTranslation(0, 16)
-				tileData := curRegion.GetTileData(int32(shadow.Style), int32(shadow.Sequence), d2enum.Floor)
-				tileSubAttrs := d2dt1.SubTileFlags{}
-				if tileData != nil {
-					tileSubAttrs = *tileData.GetSubTileFlags(subtileX, subtileY)
-				}
-				screen.DrawText("Wall %v: [HID:%t] %s", idx, shadow.Hidden, tileSubAttrs.DebugString())
-
-			}
-			screen.PushTranslation(-16, 0)
-			popN += 3
-		}
-		screen.PopN(popN)
-	}
+	//screenX, screenY := d2render.GetCursorPos()
+	//worldX, worldY := met.mapRenderer.ScreenToWorld(screenX, screenY)
+	//
+	//levelFilesToPick := make([]string, 0)
+	//fileIndex := met.fileIndex
+	//levelPreset := curRegion.LevelPreset()
+	//regionPath := curRegion.RegionPath()
+	//for n, fileRecord := range levelPreset.Files {
+	//	if len(fileRecord) == 0 || fileRecord == "" || fileRecord == "0" {
+	//		continue
+	//	}
+	//	levelFilesToPick = append(levelFilesToPick, fileRecord)
+	//	if fileRecord == regionPath {
+	//		fileIndex = n
+	//	}
+	//}
+	//if met.fileIndex == -1 {
+	//	met.fileIndex = fileIndex
+	//}
+	//met.filesCount = len(levelFilesToPick)
+	//
+	//tileX := int(math.Floor(worldX))
+	//tileY := int(math.Floor(worldY))
+	//
+	//subtileX := int((worldX - float64(int(worldX))) * 5)
+	//subtileY := int((worldY - float64(int(worldY))) * 5)
+	//
+	//regionWidth, regionHeight := curRegion.GetTileSize()
+	//if tileX >= 0 && tileY >= 0 && tileX < regionWidth && tileY < regionHeight {
+	//	tile := curRegion.Tile(tileX, tileY)
+	//	screen.PushTranslation(5, 5)
+	//	screen.DrawText("%d, %d (Tile %d.%d, %d.%d)", screenX, screenY, tileX, subtileX, tileY, subtileY)
+	//	screen.PushTranslation(0, 16)
+	//	screen.DrawText("Map: " + curRegion.LevelType().Name)
+	//	screen.PushTranslation(0, 16)
+	//	screen.DrawText("%v: %v/%v [%v, %v]", regionPath, fileIndex+1, met.filesCount, met.currentRegion, met.levelPreset)
+	//	screen.PushTranslation(0, 16)
+	//	screen.DrawText("N - next region, P - previous region")
+	//	screen.PushTranslation(0, 16)
+	//	screen.DrawText("Shift+N - next preset, Shift+P - previous preset")
+	//	screen.PushTranslation(0, 16)
+	//	screen.DrawText("Ctrl+N - next file, Ctrl+P - previous file")
+	//	screen.PushTranslation(0, 16)
+	//	popN := 7
+	//	if len(tile.Floors) > 0 {
+	//		screen.PushTranslation(0, 16)
+	//		screen.DrawText("Floors:")
+	//		screen.PushTranslation(16, 0)
+	//		for idx, floor := range tile.Floors {
+	//			popN++
+	//			screen.PushTranslation(0, 16)
+	//			tileData := curRegion.TileData(int32(floor.Style), int32(floor.Sequence), d2enum.Floor)
+	//			tileSubAttrs := d2dt1.SubTileFlags{}
+	//			if tileData != nil {
+	//				tileSubAttrs = *tileData.GetSubTileFlags(subtileX, subtileY)
+	//			}
+	//			screen.DrawText("Floor %v: [ANI:%t] %s", idx, floor.Animated, tileSubAttrs.DebugString())
+	//
+	//		}
+	//		screen.PushTranslation(-16, 0)
+	//		popN += 3
+	//	}
+	//	if len(tile.Walls) > 0 {
+	//		screen.PushTranslation(0, 16)
+	//		screen.DrawText("Walls:")
+	//		screen.PushTranslation(16, 0)
+	//		for idx, wall := range tile.Walls {
+	//			popN++
+	//			screen.PushTranslation(0, 16)
+	//			tileData := curRegion.TileData(int32(wall.Style), int32(wall.Sequence), d2enum.Floor)
+	//			tileSubAttrs := d2dt1.SubTileFlags{}
+	//			if tileData != nil {
+	//				tileSubAttrs = *tileData.GetSubTileFlags(subtileX, subtileY)
+	//			}
+	//			screen.DrawText("Wall %v: [HID:%t] %s", idx, wall.Hidden, tileSubAttrs.DebugString())
+	//
+	//		}
+	//		screen.PushTranslation(-16, 0)
+	//		popN += 3
+	//	}
+	//	if len(tile.Walls) > 0 {
+	//		screen.PushTranslation(0, 16)
+	//		screen.DrawText("Shadows:")
+	//		screen.PushTranslation(16, 0)
+	//		for idx, shadow := range tile.Shadows {
+	//			popN++
+	//			screen.PushTranslation(0, 16)
+	//			tileData := curRegion.TileData(int32(shadow.Style), int32(shadow.Sequence), d2enum.Floor)
+	//			tileSubAttrs := d2dt1.SubTileFlags{}
+	//			if tileData != nil {
+	//				tileSubAttrs = *tileData.GetSubTileFlags(subtileX, subtileY)
+	//			}
+	//			screen.DrawText("Wall %v: [HID:%t] %s", idx, shadow.Hidden, tileSubAttrs.DebugString())
+	//
+	//		}
+	//		screen.PushTranslation(-16, 0)
+	//		popN += 3
+	//	}
+	//	screen.PopN(popN)
+	//}
 
 	return nil
 }
 
 func (met *MapEngineTest) Advance(tickTime float64) error {
 	met.mapEngine.Advance(tickTime)
+	met.mapRenderer.Advance(tickTime)
 	return nil
 }
