+++ d2game/d2player/game_controls.go
@@ -6,7 +6,9 @@ import (
 	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2resource"
 	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2asset"
 	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2input"
-	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2map"
+	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2map/d2mapengine"
+	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2map/d2mapentity"
+	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2map/d2maprenderer"
 	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2render"
 	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2term"
 	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2ui"
@@ -23,9 +25,9 @@ type Panel interface {
 var missileID = 59
 
 type GameControls struct {
-	hero          *d2map.Player
-	mapEngine     *d2map.MapEngine
-	mapRenderer   *d2map.MapRenderer
+	hero          *d2mapentity.Player
+	mapEngine     *d2mapengine.MapEngine
+	mapRenderer   *d2maprenderer.MapRenderer
 	inventory     *Inventory
 	heroStats     *HeroStats
 	inputListener InputCallbackListener
@@ -37,7 +39,7 @@ type GameControls struct {
 	skillIcon   *d2ui.Sprite
 }
 
-func NewGameControls(hero *d2map.Player, mapEngine *d2map.MapEngine, mapRenderer *d2map.MapRenderer, inputListener InputCallbackListener) *GameControls {
+func NewGameControls(hero *d2mapentity.Player, mapEngine *d2mapengine.MapEngine, mapRenderer *d2maprenderer.MapRenderer, inputListener InputCallbackListener) *GameControls {
 	d2term.BindAction("setmissile", "set missile id to summon on right click", func(id int) {
 		missileID = id
 	})
@@ -76,7 +78,7 @@ func (g *GameControls) OnMouseButtonDown(event d2input.MouseEvent) bool {
 	}
 
 	if event.Button == d2input.MouseButtonRight {
-		missile, err := d2map.CreateMissile(
+		missile, err := d2mapentity.CreateMissile(
 			int(g.hero.AnimatedComposite.LocationX),
 			int(g.hero.AnimatedComposite.LocationY),
 			d2datadict.Missiles[missileID],
