+++ d2networking/d2server/game_server.go
@@ -9,9 +9,13 @@ import (
 	"log"
 	"net"
 	"strings"
+	"time"
+
+	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2map/d2mapgen"
+
+	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2map/d2mapengine"
 
 	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2enum"
-	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2map"
 	"github.com/OpenDiablo2/OpenDiablo2/d2networking/d2netpacket"
 	"github.com/OpenDiablo2/OpenDiablo2/d2networking/d2netpacket/d2netpackettype"
 	"github.com/OpenDiablo2/OpenDiablo2/d2networking/d2server/d2udpclientconnection"
@@ -21,7 +25,7 @@ import (
 
 type GameServer struct {
 	clientConnections map[string]ClientConnection
-	mapEngines        []*d2map.MapEngine
+	mapEngines        []*d2mapengine.MapEngine
 	scriptEngine      *d2script.ScriptEngine
 	udpConnection     *net.UDPConn
 	seed              int64
@@ -38,13 +42,14 @@ func Create(openNetworkServer bool) {
 
 	singletonServer = &GameServer{
 		clientConnections: make(map[string]ClientConnection),
-		mapEngines:        make([]*d2map.MapEngine, 0),
+		mapEngines:        make([]*d2mapengine.MapEngine, 0),
 		scriptEngine:      d2script.CreateScriptEngine(),
-		seed:              1592539977884044000, //time.Now().UnixNano(),
+		seed:              time.Now().UnixNano(),
 	}
 
-	mapEngine := d2map.CreateMapEngine(singletonServer.seed)
-	mapEngine.GenerateAct1Overworld(false)
+	mapEngine := d2mapengine.CreateMapEngine()
+	mapEngine.ResetMap(0, d2enum.RegionAct1Town, 100, 100) // TODO: Mapgen - Needs levels.txt stuff
+	d2mapgen.GenerateAct1Overworld(mapEngine)
 	singletonServer.mapEngines = append(singletonServer.mapEngines, mapEngine)
 
 	singletonServer.scriptEngine.AddFunction("getMapEngines", func(call otto.FunctionCall) otto.Value {
